Small edit for patched talkgroups
When 2 smartnet talkgroups are patched, they create a new talkgroup with a new number. Since any talkgroups can be patched, it makes it difficult to keep track of them. This small edit ignores the status bit of all talkgroup hex IDs, identifying them as their original talkgroup even if they are patched.